ボーン変形パラメータのデータ構造をシェーダで処理しやすいように変更 その他警告を取り除いた

+++ IQByDirectX12/IQByDirectX12/IndexBuffer.cpp
@@ -71,7 +71,7 @@ std::shared_ptr<IndexBuffer> IndexBuffer::Create(ComPtr<ID3D12Device> device, vo
 			return nullptr;
 		}
 		indexBuffer->mIndexBufferView.BufferLocation = indexBuffer->mIndexBuffer->GetGPUVirtualAddress();
-		indexBuffer->mIndexBufferView.SizeInBytes = indexSize * indexCount;
+		indexBuffer->mIndexBufferView.SizeInBytes = (UINT)(indexSize * indexCount);
 		indexBuffer->mIndexBufferView.Format = indexFormat;		
 	}
